app/requirements.txt
argon2-cffi
fastapi
lxml
psycopg2-binary
python-jose
python-multipart
sqlalchemy
uvicorn
yfinance